Fix avoidStruts
Problem: avoidStruts suddenly stoped working.
Fix: base config on desktopConfig insted of defaultConfig.

Avoid struts is a feature which avoids overlapping of ordinary windows with
docks (xmobar and trayer). For some reason it suddenly stopped working
(in the middle of ordinary desktop session, have no idea what triggered
that to happen).

According to https://github.com/xmonad/xmonad/issues/18 one should add

manageDocks in the manageHook
avoidStruts in the layoutHook
docksEventHook in the handleEventHook
docksStartupHook in the startupHook

We forgot to add 2 out of 4. But there is desktopConfig which configures
avoidStruts together with ewm (required for activity tracking software
to work).